% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeR}
\alias{applyEdgeR}
\title{applyEdgeR}
\usage{
applyEdgeR(counts, design.matrix, factors.column = NULL,
  weight.columns = NULL, contrasts = NULL, useIntercept = FALSE,
  p.threshold = 1, is.normalized = FALSE, verbose = FALSE)
}
\arguments{
\item{p.threshold}{use 1 if you want all the results to be returned}

\item{verbose}{}
}
\description{
applyEdgeR
}
