% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeRContrast}
\alias{applyEdgeRContrast}
\title{applyEdgeRContrast}
\usage{
applyEdgeRContrast(contrast, design, fit, p.threshold = 1, verbose = FALSE)
}
\arguments{
\item{p.threshold}{using 1 to be returned all the genes results}

\item{verbose}{}
}
\description{
applyEdgeRContrast
}
