% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeRGLMFit}
\alias{applyEdgeRGLMFit}
\title{applyEdgeRGLMFit}
\usage{
applyEdgeRGLMFit(
  counts,
  factors,
  design,
  is.normalized = FALSE,
  method = "TMM",
  verbose = FALSE
)
}
\arguments{
\item{is.normalized}{}
}
\description{
applyEdgeRGLMFit
}
