% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeRLRT}
\alias{applyEdgeRLRT}
\title{applyEdgeRLRT}
\usage{
applyEdgeRLRT(
  fit,
  interaction.matrix,
  interaction.term = NULL,
  p.threshold = 1,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
applyEdgeRLRT
}
