% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeRQLFit}
\alias{applyEdgeRQLFit}
\title{applyEdgeRQLFit}
\usage{
applyEdgeRQLFit(counts, factors, design, edgeroffs = NULL, verbose = FALSE)
}
\arguments{
\item{counts}{matrix of counts of features on rows and samples on columns}

\item{factors}{grouping factors vector of the same length of the counts column}

\item{design}{a design matrix typically created with \link[stats]{model.matrix}}

\item{edgeroffs}{offset for \link[edger]{glmQLFit}}

\item{verbose}{}
}
\value{
a DGEGLM class object
}
\description{
creates a DGEList object, estimate dispersion and applies
glmQLFit
}
\examples{
TBD
}
