% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{attachGeneColumnToDf}
\alias{attachGeneColumnToDf}
\title{attachGeneColumnToDf}
\usage{
attachGeneColumnToDf(
  mainDf,
  genesMap,
  rowNamesIdentifier = c("entrezgene", "ensembl", "symbol", "ENTREZID", "SYMBOL"),
  mapFromIdentifier = NULL,
  mapToIdentifier = NULL
)
}
\arguments{
\item{mapToIdentifier}{}
}
\description{
attachGeneColumnToDf
}
