% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{attachMeans}
\alias{attachMeans}
\title{attachMeans}
\usage{
attachMeans(
  normalized.counts,
  design.matrix,
  factor.column,
  contrast.name,
  de.results
)
}
\arguments{
\item{de.results}{}
}
\description{
attachMeans
}
