% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFunctions.R
\name{computeGeneMeansOverGroups}
\alias{computeGeneMeansOverGroups}
\title{computeGeneMeansOverGroups}
\usage{
computeGeneMeansOverGroups(counts, design, groupColumn)
}
\arguments{
\item{groupColumn}{}
}
\value{

}
\description{
computeGeneMeansOverGroups
}
