% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{computeMeans}
\alias{computeMeans}
\title{computeMeans}
\usage{
computeMeans(counts, design.matrix, factors.column, contrst, genes)
}
\arguments{
\item{genes}{}
}
\value{

}
\description{
computeMeans
}
