% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{constructInteractionMatrix}
\alias{constructInteractionMatrix}
\title{constructInteractionMatrix}
\usage{
constructInteractionMatrix(
  design.matrix,
  genotype.col,
  genotype.ref = NULL,
  condition.col,
  weight.col,
  weights = NULL,
  useIntercept = TRUE
)
}
\arguments{
\item{design.matrix}{a design matrix of the experiment where each row
describes a sample}

\item{genotype.col}{the column name/index describing the genotype in
the design matrix}

\item{genotype.ref}{an optional character indicating which factor name to
use as genotype reference}

\item{condition.col}{the column name/index describing the condition in
the design matrix}

\item{weight.col}{optional column name/index describing the weigths in
the design matrix (see also weigths parameter)}

\item{weights}{optional matrix with weights to use in the
interaction matrix (see also weigth.col parameter)}

\item{useIntercept}{}
}
\value{
a design matrix with interaction terms
}
\description{
constructInteractionMatrix
}
