% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convertGenesViaBiomart}
\alias{convertGenesViaBiomart}
\title{convertGenesViaBiomart}
\usage{
convertGenesViaBiomart(
  specie = c("hg38", "mm10", "rnor6"),
  attrs = NULL,
  filter = NULL,
  filter.values = NULL
)
}
\arguments{
\item{filter.values}{}
}
\description{
convertGenesViaBiomart
}
