% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convertGenesViaMouseDb}
\alias{convertGenesViaMouseDb}
\title{convertGenesViaMouseDb}
\usage{
convertGenesViaMouseDb(gene.list, fromType = c("SYMBOL", "ENTREZID"),
  toType = c("SYMBOL", "ENTREZID"))
}
\arguments{
\item{toType}{}
}
\description{
convertGenesViaMouseDb
}
