% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlotFunctions.R,
%   R/luciaVolcanoFunctions.R
\name{createSignLabelsNumbers}
\alias{createSignLabelsNumbers}
\title{Title}
\usage{
createSignLabelsNumbers(proc.df, column.name)

createSignLabelsNumbers(proc.df, column.name)
}
\arguments{
\item{column.name}{}
}
\description{
Title

createSignLabelsNumbers
}
