% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalFunctions.R
\name{enrichGOGProfiler}
\alias{enrichGOGProfiler}
\title{Title}
\usage{
enrichGOGProfiler(
  gene.list.to.enrich,
  organism.name = c("rnorvegicus", "mmusculus", "hsapiens"),
  exclude.electr.flag = FALSE,
  ontology = c("BP", "MF", "CC"),
  significant.flag = FALSE,
  min.isect.size = 0,
  correction.method = "fdr",
  functional.folder,
  filename
)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Title
}
