% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalFunctions.R
\name{enrichPathwayGProfiler}
\alias{enrichPathwayGProfiler}
\title{Title}
\usage{
enrichPathwayGProfiler(
  gene.list.to.enrich,
  organism.name = c("rnorvegicus", "mmusculus", "hsapiens"),
  exclude.electr.flag = FALSE,
  path.label = c("KEGG", "REAC"),
  significant.flag = FALSE,
  correction.method = "fdr",
  functional.folder,
  filename
)
}
\arguments{
\item{filename}{}
}
\description{
Title
}
