% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalFunctions.R
\name{enrichSuitedSingleList}
\alias{enrichSuitedSingleList}
\title{Title}
\usage{
enrichSuitedSingleList(
  de.gene.list,
  functional.folder,
  filename,
  organism = "hsapiens"
)
}
\arguments{
\item{organism}{}
}
\description{
Title
}
