% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLowCounts.R
\name{filterLowCounts}
\alias{filterLowCounts}
\title{filterLowCounts}
\usage{
filterLowCounts(counts.dataframe, is.normalized = c(TRUE, FALSE),
  design.dataframe, cond.col.name = NULL, method.type = c("CPM", "Wilcoxon",
  "Proportion"), cv.percentage = 100, cpm.cutoff = 1, seq.depth = NULL,
  verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
filterLowCounts
}
