% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalHistPlotFunctions.R
\name{generateGGHist}
\alias{generateGGHist}
\title{generateGGHist}
\usage{
generateGGHist(processed.de.results, strings, nbins = 30)
}
\arguments{
\item{nbins}{}
}
\value{

}
\description{
generateGGHist
}
