% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlotFunctions.R
\name{GenerateGGMA}
\alias{GenerateGGMA}
\title{GenerateGGMA}
\usage{
GenerateGGMA(processed.de.results, strings, plotly.flag = FALSE)
}
\arguments{
\item{processed.de.results}{dataframe. D-E results processed with ProcessDEResultsForPlot}

\item{strings}{character list. Returned by GeneratePlotStrings}

\item{plotly.flag}{logical. Is it a ggplotly object?}
}
\value{
none
}
\description{
Generate an MA-Plot using ggplot
}
