% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalVariables.R, R/plotUtils.R
\name{GeneratePlotStrings}
\alias{GeneratePlotStrings}
\title{Title}
\usage{
GeneratePlotStrings(path = NULL, prefix, plot.type)

GeneratePlotStrings(path = NULL, prefix, plot.type)
}
\arguments{
\item{plot.type}{}
}
\value{



}
\description{
Title

GeneratePlotStrings
}
