% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luciaVolcanoFunctions.R
\name{luciaVolcanoPlot}
\alias{luciaVolcanoPlot}
\title{luciaVolcanoPlot}
\usage{
luciaVolcanoPlot(res.o, positive.controls.df, prefix, threshold = 0.01)
}
\arguments{
\item{threshold}{}
}
\description{
luciaVolcanoPlot
}
