% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{NormalizeData}
\alias{NormalizeData}
\title{NormalizeData}
\usage{
NormalizeData(
  data.to.normalize,
  norm.type = c("fqua", "uqua", "tmm", "ruvg", "ruvs"),
  design.matrix = NULL,
  factors.column = NULL,
  estimated.genes = NULL,
  is.log = FALSE,
  ruv_k = 1
)
}
\arguments{
\item{is.log}{}
}
\description{
NormalizeData
}
