% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{NormalizeDataSE}
\alias{NormalizeDataSE}
\title{NormalizeDataSE}
\usage{
NormalizeDataSE(
  se,
  norm.type = c("fqua", "uqua", "tmm", "ruvg", "ruvs"),
  factors.column = NULL,
  control.genes = NULL,
  assay = "counts",
  is.log = FALSE,
  ruv_k = 1
)
}
\arguments{
\item{is.log}{}
}
\description{
NormalizeDataSE
}
