% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{PlotCountsAlongTimes}
\alias{PlotCountsAlongTimes}
\title{PlotCountsAlongTimes}
\usage{
PlotCountsAlongTimes(normalized.counts, design.matrix, gene.name,
  gene.name.column.name = "gene.names", show.plot.flag = TRUE,
  plotly.flag = FALSE, save.plot = FALSE, plot.folder = NULL,
  prefix.plot = "Counts Plot")
}
\arguments{
\item{prefix.plot}{}
}
\description{
PlotCountsAlongTimes
}
