% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlotFunctions.R
\name{PlotMAPlotCounts}
\alias{PlotMAPlotCounts}
\title{PlotMAPlotCounts}
\usage{
PlotMAPlotCounts(
  de.results,
  counts.dataframe,
  design.matrix,
  show.plot.flag = TRUE,
  plotly.flag = FALSE,
  save.plot = FALSE,
  plot.folder = NULL,
  prefix.plot = "MA_Plot",
  threshold = 0.05
)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
PlotMAPlotCounts
}
