% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlotFunctions.R
\name{PlotVolcanoPlot}
\alias{PlotVolcanoPlot}
\title{Title}
\usage{
PlotVolcanoPlot(
  de.results,
  counts.dataframe = NULL,
  design.matrix = NULL,
  show.plot.flag = TRUE,
  plotly.flag = FALSE,
  save.plot = FALSE,
  plot.folder = NULL,
  prefix.plot = NULL,
  threshold,
  positive.ctrls.list = NULL,
  color.palette = c("blue2", "red2")
)
}
\arguments{
\item{positive.ctrls.list}{}
}
\description{
Title
}
