% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{ProcessDEResultsForPlot}
\alias{ProcessDEResultsForPlot}
\title{ProcessDEResultsForPlot}
\usage{
ProcessDEResultsForPlot(
  de.results,
  threshold = 0.05,
  counts.dataframe = NULL,
  design.matrix = NULL,
  pos.ctrls.list = NULL
)
}
\arguments{
\item{pos.ctrls.list}{}
}
\value{

}
\description{
ProcessDEResultsForPlot
}
