% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizationFunctions.R
\name{RUVgNormalizationFunction}
\alias{RUVgNormalizationFunction}
\title{RUVgNormalizationFunction}
\usage{
RUVgNormalizationFunction(data.to.normalize, design.matrix, desMatColStr,
  estimated.gene.names, k = 1, isLog = FALSE, gene.names.col = NULL)
}
\arguments{
\item{gene.names.col}{}
}
\description{
RUVgNormalizationFunction
}
