% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn3.R
\name{SaveInteserctionsList}
\alias{SaveInteserctionsList}
\title{SaveInteserctionsList}
\usage{
SaveInteserctionsList(
  gene.list,
  conversion.map,
  root.dir,
  prefix,
  labels.list,
  enrich.lists.flag = FALSE,
  heatmap.flag = FALSE,
  expression.data = NULL
)
}
\arguments{
\item{expression.data}{}
}
\description{
SaveInteserctionsList
}
