% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{SelectGenesFromHistBreaks}
\alias{SelectGenesFromHistBreaks}
\title{SelectGenesFromHistBreaks}
\usage{
SelectGenesFromHistBreaks(data.counts, breakss = 100, n.genes.per.break = 1)
}
\arguments{
\item{n.genes.per.break}{}
}
\description{
SelectGenesFromHistBreaks
}
