% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subsetDfByCol}
\alias{subsetDfByCol}
\title{subsetDfByCol}
\usage{
subsetDfByCol(df, list, colname = NULL)
}
\arguments{
\item{df}{a dataframe.}

\item{list}{a list of identifiers.}

\item{colname}{the df column where to check the list in the df.
If NULL, the rownames will be used.}
}
\description{
subsetDfByCol
}
