% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{TailDeGenesByPAdj}
\alias{TailDeGenesByPAdj}
\title{TailDeGenesByPAdj}
\usage{
TailDeGenesByPAdj(de.genes, n.genes = 10, threshold = 0.05)
}
\arguments{
\item{threshold}{}
}
\description{
TailDeGenesByPAdj
}
