% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalVariables.R, R/plotUtils.R
\name{UpdatePrefix}
\alias{UpdatePrefix}
\title{Title}
\usage{
UpdatePrefix(prefix, ...)

UpdatePrefix(prefix, ...)
}
\arguments{
\item{...}{}
}
\value{



}
\description{
Title

UpdatePrefix
}
