% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn2.R
\name{Venn2de}
\alias{Venn2de}
\title{Title}
\usage{
Venn2de(
  x,
  y,
  label1,
  label2,
  title = NULL,
  plot.dir,
  conversion.map = NULL,
  intersection.flag = FALSE,
  enrich.lists.flag = FALSE,
  fill_color = c("darkcyan", "darkcyan"),
  fill_alpha = 0.4,
  stroke_size = 0.5,
  stroke_alpha = 0.5,
  stroke_color = "black",
  set_name_size = 6,
  prefix = "",
  plot.heatmap = FALSE
)
}
\arguments{
\item{plot.heatmap}{}
}
\description{
Title
}
