% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn3.R
\name{Venn3de}
\alias{Venn3de}
\title{Venn3de}
\usage{
Venn3de(
  x,
  y,
  z,
  label1 = "x",
  label2 = "y",
  label3 = "z",
  title = "Venn Diagram",
  intersection.flag = TRUE,
  intersection.exclusion.flag = FALSE,
  plot.dir = NULL,
  enrich.lists.flag = FALSE,
  conversion.map = NULL,
  plot.heatmap = FALSE,
  expression.data.list = NULL,
  expression.colname = "logFC"
)
}
\arguments{
\item{x}{a list of elements}

\item{y}{a list of elements}

\item{z}{a list of elements}

\item{label1}{the label for the x list}

\item{label2}{the label for the y list}

\item{label3}{the label for the z list}

\item{title}{an optional title for the Venn plot}

\item{intersection.flag}{a boolean for saving the lists of intersected
elements}

\item{intersection.exclusion.flag}{a boolean for saving the lists of unique
sets lists}

\item{plot.dir}{the directory to save the intersections}

\item{enrich.lists.flag}{If to perform the functional enrichment for the
produced lists}

\item{conversion.map}{a data.frame with gene ids conversions
(i.e. Ensembl_id, Gene_name)}

\item{plot.heatmap}{flag indicating if plotting a heatmap for the main
intersection}

\item{expression.data.list}{a list of dataframe/matrix with rownames with
expression values ordered for x,y,z}

\item{expression.colname}{the name of the column within the expression values}
}
\description{
TBW
}
\examples{
TBD
}
