% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeR}
\alias{applyEdgeR}
\title{applyEdgeR}
\usage{
applyEdgeR(
  counts,
  design.matrix,
  factors.column = NULL,
  weight.columns = NULL,
  contrasts = NULL,
  useIntercept = FALSE,
  p.threshold = 1,
  is.normalized = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{factors.column}{a character string indicating where to find the contrasts into
the design.matrix argument}

\item{weight.columns}{a character indicating the colnames of the weights to
add to the model.matrix function}

\item{contrasts}{a character indicating the contrasts to test (i.e. "c1-c2"),
more than one can be passed as a list}

\item{p.threshold}{numeric cutoff for adjusted p-values (default is 1)}

\item{is.normalized}{a logical indicating if the counts are normalized}

\item{verbose}{logical for printing additional messages}
}
\description{
applyEdgeR
}
