% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRFunctions.R
\name{applyEdgeRQLFit}
\alias{applyEdgeRQLFit}
\title{applyEdgeRQLFit}
\usage{
applyEdgeRQLFit(
  counts,
  factors,
  design,
  is.normalized = FALSE,
  method = "TMM",
  verbose = FALSE
)
}
\arguments{
\item{is.normalized}{}
}
\value{

}
\description{
applyEdgeRQLFit
}
