% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luciaVolcanoFunctions.R
\name{createSignLabelsNumbers}
\alias{createSignLabelsNumbers}
\title{createSignLabelsNumbers}
\usage{
createSignLabelsNumbers(proc.df, column.name)
}
\arguments{
\item{column.name}{}
}
\description{
createSignLabelsNumbers
}
