% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizationFunctions.R
\name{EstimateNegativeControlGenesForRUV}
\alias{EstimateNegativeControlGenesForRUV}
\title{EstimateNegativeControlGenesForRUV}
\usage{
EstimateNegativeControlGenesForRUV(de.genes, n.tail.genes = 2000,
  counts.dataset, n.genes.per.hist.break = 1, threshold = 0.05)
}
\arguments{
\item{threshold}{}
}
\description{
EstimateNegativeControlGenesForRUV
}
