% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLowCounts.R
\name{filterLowCounts}
\alias{filterLowCounts}
\title{filterLowCounts}
\usage{
filterLowCounts(
  se,
  is.normalized = c(TRUE, FALSE),
  cond.col.name = NULL,
  method.type = c("CPM", "Wilcoxon", "Proportion"),
  cv.percentage = 100,
  cpm.cutoff = 1,
  seq.depth = NULL,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
filterLowCounts
}
