% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{GeneratePlotStrings}
\alias{GeneratePlotStrings}
\title{GeneratePlotStrings}
\usage{
GeneratePlotStrings(path = NULL, prefix, plot.type)
}
\arguments{
\item{plot.type}{}
}
\description{
GeneratePlotStrings
}
