% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizationFunctions.R
\name{NormalizeData}
\alias{NormalizeData}
\title{NormalizeData}
\usage{
NormalizeData(
  data.to.normalize,
  norm.type = c("fqua", "uqua", "tmm", "ruvg"),
  design.matrix = NULL,
  design.matrix.factors.column = NULL,
  estimated.genes = NULL,
  is.log = FALSE
)
}
\arguments{
\item{is.log}{}
}
\value{

}
\description{
NormalizeData
}
