% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn2.R
\name{Venn2de}
\alias{Venn2de}
\title{Title}
\usage{
Venn2de(
  x,
  y,
  label1,
  label2,
  title,
  plot.dir,
  conversion.map = NULL,
  intersection.flag = FALSE,
  enrich.lists.flag = FALSE,
  prefix = "",
  plot.heatmap = FALSE
)
}
\arguments{
\item{plot.heatmap}{}
}
\description{
Title
}
