% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{WriteDataFrameAsTsv}
\alias{WriteDataFrameAsTsv}
\title{WriteDataFrameAsXls}
\usage{
WriteDataFrameAsTsv(
  data.frame.to.save,
  file.name.path,
  col.names = TRUE,
  row.names = TRUE
)
}
\arguments{
\item{col.names}{}
}
\description{
WriteDataFrameAsXls
}
